+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 ## Bug Fixes:
 - Awakened Spellbook
 - Drake Companion
+- Item Updater / Configuration
 - Meta Magic
 - Primal Companion
-
+  
 # 0.5.28 Release:
 ## New Content:
 - Sorcerer Metamagic: Careful Spell, Twinned Spell, Transmuted Spell, Seeking Spell, Heightened Spell, and Empowered Spell (Thanks Autumn)